doc(atomic): dispatch event for Atomic production releases
https://coveord.atlassian.net/browse/KIT-858
